more code converted

+++ Weapons.cs
@@ -4,33 +4,33 @@ namespace RustEQ;
 
 internal static class Weapons
 {
-        public const int NUM_ATTACHMENTS = 4;
-        public const int NUM_WEAPONS = 13;
+    public const int NUM_ATTACHMENTS = 4;
+    public const int NUM_WEAPONS = 13;
 
-        public const byte ATTACHMENT_NONE = 0;
-        public const byte ATTACHMENT_HOLO = 1;
-        public const byte ATTACHMENT_SCOPE8 = 2;
-        public const byte ATTACHMENT_SCOPE16 = 3;
+    public const byte ATTACHMENT_NONE = 0;
+    public const byte ATTACHMENT_HOLO = 1;
+    public const byte ATTACHMENT_SCOPE8 = 2;
+    public const byte ATTACHMENT_SCOPE16 = 3;
 
-        public const byte WEAPON_M249 = 0;
-        public const byte WEAPON_AK47 = 1;
-        public const byte WEAPON_LIGHTRIFLE = 2;
-        public const byte WEAPON_M39 = 3;
-        public const byte WEAPON_SEMIRIFLE = 4;
-        public const byte WEAPON_MP5 = 5;
-        public const byte WEAPON_CUSTOMSMG = 6;
-        public const byte WEAPON_THOMPSON = 7;
-        public const byte WEAPON_PYTHONREVOLVER = 8;
-        public const byte WEAPON_M92 = 9;
-        public const byte WEAPON_REVOLVER = 10;
-        public const byte WEAPON_SEMIPISTOL = 11;
-        public const byte WEAPON_NAILGUN = 12;
+    public const byte WEAPON_M249 = 0;
+    public const byte WEAPON_AK47 = 1;
+    public const byte WEAPON_LIGHTRIFLE = 2;
+    public const byte WEAPON_M39 = 3;
+    public const byte WEAPON_SEMIRIFLE = 4;
+    public const byte WEAPON_MP5 = 5;
+    public const byte WEAPON_CUSTOMSMG = 6;
+    public const byte WEAPON_THOMPSON = 7;
+    public const byte WEAPON_PYTHONREVOLVER = 8;
+    public const byte WEAPON_M92 = 9;
+    public const byte WEAPON_REVOLVER = 10;
+    public const byte WEAPON_SEMIPISTOL = 11;
+    public const byte WEAPON_NAILGUN = 12;
 
-        public static readonly string[] AttachmentStringMap = new string[NUM_ATTACHMENTS] { "No Scope", "Holosight", "Scope x8", "Scope x16" };
-        public static readonly string[] WeaponStringMap = new string[NUM_WEAPONS] { "M249", "Ak-47", "Light-Rifle", "M39", "Semi-AR", "Mp5", "Custom-Smg", "Thompson", "El Python", "M92", "Revolver", "Semi-Pistol", "Nailgun" };
+    public static readonly string[] AttachmentStringMap = new string[NUM_ATTACHMENTS] { "No Scope", "Holosight", "Scope x8", "Scope x16" };
+    public static readonly string[] WeaponStringMap = new string[NUM_WEAPONS] { "M249", "Ak-47", "Light-Rifle", "M39", "Semi-AR", "Mp5", "Custom-Smg", "Thompson", "El Python", "M92", "Revolver", "Semi-Pistol", "Nailgun" };
 
-        public static readonly float[] WEAPON_DEFAULT_ADSZOOM = new float[NUM_WEAPONS]
-        {
+    public static readonly float[] WEAPON_DEFAULT_ADSZOOM = new float[NUM_WEAPONS]
+    {
             5.0f / 3.0f,
             5.0f / 3.0f,
             5.0f / 3.0f,
@@ -44,26 +44,26 @@ internal static class Weapons
             15.0f / 13.0f,
             15.0f / 13.0f,
             5.0f / 4.0f
-        };
+    };
 
-        public static readonly float[] ATTACHMENT_RCS_MULTIPLIER = new float[NUM_ATTACHMENTS]
-        {
+    public static readonly float[] ATTACHMENT_RCS_MULTIPLIER = new float[NUM_ATTACHMENTS]
+    {
             1.0f,
             1.0f,
             0.8f,
             0.8f
-        };
+    };
 
-        public static readonly float[] ATTACHMENT_ZOOM = new float[NUM_ATTACHMENTS]
-        {
+    public static readonly float[] ATTACHMENT_ZOOM = new float[NUM_ATTACHMENTS]
+    {
             1.0f,
             2.0f,
             8.0f,
             16.0f
-        };
+    };
 
-        public static readonly bool[] WEAPON_ISCURVEBASED = new bool[NUM_WEAPONS]
-        {
+    public static readonly bool[] WEAPON_ISCURVEBASED = new bool[NUM_WEAPONS]
+    {
             false,
             true,
             true,
@@ -77,10 +77,10 @@ internal static class Weapons
             false,
             false,
             false
-        };
+    };
 
-        public static readonly AnimationCurvePair[] WEAPON_ANIMATIONCURVE = new AnimationCurvePair[NUM_WEAPONS]
-        {
+    public static readonly AnimationCurvePair[] WEAPON_ANIMATIONCURVE = new AnimationCurvePair[NUM_WEAPONS]
+    {
             AnimationCurveFactory.NullCurve,
             AnimationCurveFactory.AkCurve,
             AnimationCurveFactory.LrCurve,
@@ -94,10 +94,10 @@ internal static class Weapons
             AnimationCurveFactory.NullCurve,
             AnimationCurveFactory.NullCurve,
             AnimationCurveFactory.NullCurve
-        };
+    };
 
-        public static readonly Vector2[] WEAPON_BASELINE_SCALAR = new Vector2[NUM_WEAPONS]
-        {
+    public static readonly Vector2[] WEAPON_BASELINE_SCALAR = new Vector2[NUM_WEAPONS]
+    {
             new(0, -5.5f),
             new(8, -30f),
             new(5, -12f),
@@ -111,27 +111,27 @@ internal static class Weapons
             new(0, -4.5f),
             new(0, -7f),
             new(0, -4.5f)
-        };
+    };
 
-        public static readonly float[] WEAPON_TIMEPERSHOT_MS = new float[NUM_WEAPONS]
-        {
-            60000.0f / 500.0f,
-            60000.0f / 450.0f,
-            60000.0f / 500.0f,
-            200.0f,
-            175.0f,
-            100.0f,
-            100.0f,
-            60000.0f / 462.0f,
-            150.0f,
-            150.0f,
-            175.0f,
-            150.0f,
-            150.0f
-        };
+    public static readonly long[] WEAPON_TIMEPERSHOT_MS = new long[NUM_WEAPONS]
+    {
+            60000 / 500,
+            60000 / 450,
+            60000 / 500,
+            200,
+            175,
+            100,
+            100,
+            60000 / 462,
+            150,
+            150,
+            175,
+            150,
+            150
+    };
 
-        public static readonly bool[] WEAPON_ISAUTOMATIC = new bool[NUM_WEAPONS]
-        {
+    public static readonly bool[] WEAPON_ISAUTOMATIC = new bool[NUM_WEAPONS]
+    {
             true,
             true,
             true,
@@ -145,10 +145,10 @@ internal static class Weapons
             false,
             false,
             false
-        };
+    };
 
-        public static readonly int[] WEAPON_NUM_BULLETS = new int[NUM_WEAPONS]
-        {
+    public static readonly int[] WEAPON_NUM_BULLETS = new int[NUM_WEAPONS]
+    {
             100,
             30,
             30,
@@ -162,10 +162,10 @@ internal static class Weapons
             8,
             10,
             16
-        };
+    };
 
-        public static readonly float[] WEAPON_MOVEMENT_PENALTY = new float[NUM_WEAPONS]
-        {
+    public static readonly float[] WEAPON_MOVEMENT_PENALTY = new float[NUM_WEAPONS]
+    {
             1.25f,
             0.0f,
             0.0f,
@@ -179,10 +179,10 @@ internal static class Weapons
             0.0f,
             0.5f,
             0.0f
-        };
+    };
 
-        public static readonly float[] WEAPON_ADSSCALE = new float[NUM_WEAPONS]
-        {
+    public static readonly float[] WEAPON_ADSSCALE = new float[NUM_WEAPONS]
+    {
             0.5f,
             1.0f,
             1.0f,
@@ -196,5 +196,5 @@ internal static class Weapons
             0.6f,
             0.6f,
             0.6f
-        };
+    };
 }
